chore: simplifying code, removing duplicate tests
