+++ src/__tests__/optimize.test.ts
@@ -4,7 +4,6 @@ import { Shape1D } from 'ml-peak-shape-generator';
 import { generateSpectrum } from 'spectrum-generator';
 
 import { optimize } from '../index';
-import { getSumOfShapes } from '../shapes/getSumOfShapes';
 
 expect.extend({ toBeDeepCloseTo, toMatchCloseTo });
 
@@ -15,13 +14,17 @@ for (let i = 0; i < nbPoints; i++) {
   x[i] = (i - nbPoints / 2) * xFactor;
 }
 
-describe('Optimize sum of Lorentzian', () => {
+describe('Optimize sum of Lorentzians', () => {
   const peaks = [
+    { x: -0.5, y: 1, fwhm: 0.05, shape: { kind: 'lorentzian' } },
+    { x: 0.5, y: 1, fwhm: 0.05, shape: { kind: 'lorentzian' } },
+  ];
+  const peaksGenerator = [
     { x: -0.5, y: 1, fwhm: 0.05 },
     { x: 0.5, y: 1, fwhm: 0.05 },
   ];
 
-  const data: DataXY = generateSpectrum(peaks, {
+  const data: DataXY = generateSpectrum(peaksGenerator, {
     generator: {
       from: -1,
       to: 1,
@@ -29,56 +32,22 @@ describe('Optimize sum of Lorentzian', () => {
     },
   });
 
-  it('group of two GL', () => {
-    let pTrue: number[] = [-0.5, 0.001, 0.31, 0, 0.5, 0.001, 0.31, 0];
-    let func = getSumOfShapes([
-      { x: -0.5, y: 1, fwhm: 0.05, shape: { kind: 'lorentzian' } as Shape1D },
-      { x: 0.5, y: 1, fwhm: 0.05, shape: { kind: 'lorentzian' } as Shape1D },
-    ]);
-    let yData = func(pTrue);
+  it('positive maxima peaks', () => {
     let peakList = [
       {
-        x: -0.5,
-        y: 0.0009,
-        fwhm: (xFactor * nbPoints) / 8,
+        x: -0.52,
+        y: 0.9,
+        fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
       {
         x: 0.52,
-        y: 0.0009,
-        fwhm: (xFactor * nbPoints) / 8,
-        shape: { kind: 'lorentzian' } as Shape1D,
-      },
-    ];
-    let result = optimize({ x, y: x.map((i) => yData(i)) }, peakList, {
-      shape: { kind: 'lorentzian' } as Shape1D,
-    });
-    for (let i = 0; i < 2; i++) {
-      let pFit = result.peaks[i];
-      expect(pFit.x).toBeCloseTo(pTrue[i * 4], 2);
-      expect(pFit.y).toBeCloseTo(pTrue[i * 4 + 1], 2);
-      expect(pFit.fwhm).toBeCloseTo(pTrue[i * 4 + 2], 1);
-    }
-  });
-
-  it('positive maxima peaks', () => {
-    let peakList = [
-      {
-        x: -0.5,
-        y: 1,
-        fwhm: 0.08,
-        shape: { kind: 'lorentzian' } as Shape1D,
-      },
-      {
-        x: 0.5,
-        y: 1,
+        y: 0.9,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
     ];
-    let result = optimize(data, peakList, {
-      shape: { kind: 'lorentzian' } as Shape1D,
-    });
+    let result = optimize(data, peakList);
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
       expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
@@ -90,28 +59,26 @@ describe('Optimize sum of Lorentzian', () => {
   it('baseline higher than zero', () => {
     let peakList = [
       {
-        x: -0.5,
-        y: 3,
+        x: -0.52,
+        y: 2.9,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
       {
-        x: 0.5,
-        y: 3,
+        x: 0.52,
+        y: 2.9,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
     ];
     let dataCopy = { ...data };
-    dataCopy.y.map((el) => el + 2);
+    dataCopy.y.map((el: number) => el + 2);
     const peaksCopy = [
       { x: -0.5, y: 3, fwhm: 0.05 },
       { x: 0.5, y: 3, fwhm: 0.05 },
     ];
 
-    let result = optimize(dataCopy, peakList, {
-      shape: { kind: 'lorentzian' } as Shape1D,
-    });
+    let result = optimize(dataCopy, peakList);
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
       expect(pFit.x).toBeCloseTo(peaksCopy[i].x, 1);
@@ -123,23 +90,21 @@ describe('Optimize sum of Lorentzian', () => {
   it('negative maxima peaks', () => {
     let peakList = [
       {
-        x: -0.5,
+        x: -0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
       {
-        x: 0.5,
+        x: 0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
     ];
     let modifiedData = data;
-    modifiedData.y.map((value) => value - 2);
-    let result = optimize(modifiedData, peakList, {
-      shape: { kind: 'lorentzian' } as Shape1D,
-    });
+    modifiedData.y.map((value: number) => value - 2);
+    let result = optimize(modifiedData, peakList);
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
       expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
@@ -151,23 +116,21 @@ describe('Optimize sum of Lorentzian', () => {
   it('Negative peaks', () => {
     let peakList = [
       {
-        x: -0.5,
+        x: -0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
       {
-        x: 0.5,
+        x: 0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
     ];
     let modifiedData = data;
-    modifiedData.y.map((value) => -value);
-    let result = optimize(modifiedData, peakList, {
-      shape: { kind: 'lorentzian' } as Shape1D,
-    });
+    modifiedData.y.map((value: number) => -value);
+    let result = optimize(modifiedData, peakList);
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
       expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
@@ -179,23 +142,21 @@ describe('Optimize sum of Lorentzian', () => {
   it('minima peaks', () => {
     let peakList = [
       {
-        x: -0.5,
+        x: -0.52,
         y: 0,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
       {
-        x: 0.5,
+        x: 0.52,
         y: 0,
         fwhm: 0.08,
         shape: { kind: 'lorentzian' } as Shape1D,
       },
     ];
     let modifiedData = data;
-    modifiedData.y.map((value) => 1 - value);
-    let result = optimize(modifiedData, peakList, {
-      shape: { kind: 'lorentzian' } as Shape1D,
-    });
+    modifiedData.y.map((value: number) => 1 - value);
+    let result = optimize(modifiedData, peakList);
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
       expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
@@ -207,63 +168,39 @@ describe('Optimize sum of Lorentzian', () => {
 
 describe('Optimize sum of Gaussians', () => {
   const peaks = [
-    { x: -0.5, y: 1, fwhm: 0.05, shape: { kind: 'gaussian' } as Shape1D },
-    { x: 0.5, y: 1, fwhm: 0.05, shape: { kind: 'gaussian' } as Shape1D },
+    { x: -0.5, y: 1, fwhm: 0.05, shape: { kind: 'gaussian' } },
+    { x: 0.5, y: 1, fwhm: 0.05, shape: { kind: 'gaussian' } },
   ];
-  const data: DataXY = generateSpectrum(peaks, {
+
+  const peaksGenerator = [
+    { x: -0.5, y: 1, fwhm: 0.05 },
+    { x: 0.5, y: 1, fwhm: 0.05 },
+  ];
+
+  const data: DataXY = generateSpectrum(peaksGenerator, {
     generator: {
-      from: -1,
-      to: 1,
+      from: -2,
+      to: 2,
       nbPoints: 101,
     },
-  });
-
-  it('group of two GL', () => {
-    let pTrue = [-0.5, 0.001, 0.31, 0, 0.5, 0.001, 0.31, 0];
-    let func = getSumOfShapes(peaks);
-    let yData = func(pTrue);
-    let result = optimize(
-      { x, y: x.map((i) => yData(i)) },
-      [
-        {
-          x: -0.5,
-          y: 0.0009,
-          fwhm: (xFactor * nbPoints) / 8,
-          shape: { kind: 'gaussian' } as Shape1D,
-        },
-        {
-          x: 0.52,
-          y: 0.0009,
-          fwhm: (xFactor * nbPoints) / 8,
-          shape: { kind: 'gaussian' } as Shape1D,
-        },
-      ],
-      { shape: { kind: 'gaussian' } as Shape1D },
-    );
-    for (let i = 0; i < 2; i++) {
-      let pFit = result.peaks[i];
-      expect(pFit.x).toBeCloseTo(pTrue[4 * i], 2);
-      expect(pFit.y).toBeCloseTo(pTrue[4 * i + 1], 2);
-      expect(pFit.fwhm).toBeCloseTo(pTrue[4 * i + 2], 2);
-    }
   });
 
   it('positive maxima peaks', () => {
     let peakList = [
       {
-        x: -0.5,
-        y: 1,
+        x: -0.52,
+        y: 0.9,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
       {
-        x: 0.5,
-        y: 1,
+        x: 0.52,
+        y: 0.9,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
     ];
-    let result = optimize(data, peakList, { shape: { kind: 'gaussian' } });
+    let result = optimize(data, peakList);
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
       expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
@@ -271,30 +208,32 @@ describe('Optimize sum of Gaussians', () => {
       expect(pFit.fwhm).toBeCloseTo(peaks[i].fwhm, 1);
     }
   });
-
   it('negative maxima peaks', () => {
     let peakList = [
       {
-        x: -0.5,
+        x: -0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
       {
-        x: 0.5,
+        x: 0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
     ];
     let modifiedData = data;
-    modifiedData.y.map((value) => value - 2);
+    modifiedData.y.map((value: number) => value - 2);
     let result = optimize(modifiedData, peakList, {
-      shape: { kind: 'gaussian' },
+      optimization: {
+        kind: 'lm',
+        options: { maxIterations: 500, damping: 0.5, errorTolerance: 1e-8 },
+      },
     });
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
-      expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
+      expect(pFit.x).toBeCloseTo(peaks[i].x, 0);
       expect(pFit.y).toBeCloseTo(peaks[i].y, 1);
       expect(pFit.fwhm).toBeCloseTo(peaks[i].fwhm, 1);
     }
@@ -303,26 +242,29 @@ describe('Optimize sum of Gaussians', () => {
   it('Negative peaks', () => {
     let peakList = [
       {
-        x: -0.5,
+        x: -0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
       {
-        x: 0.5,
+        x: 0.52,
         y: -1,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
     ];
     let modifiedData = data;
-    modifiedData.y.map((value) => -value);
+    modifiedData.y.map((value: number) => -value);
     let result = optimize(modifiedData, peakList, {
-      shape: { kind: 'gaussian' },
+      optimization: {
+        kind: 'lm',
+        options: { maxIterations: 500, damping: 0.5, errorTolerance: 1e-8 },
+      },
     });
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
-      expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
+      expect(pFit.x).toBeCloseTo(peaks[i].x, 0);
       expect(pFit.y).toBeCloseTo(peaks[i].y, 1);
       expect(pFit.fwhm).toBeCloseTo(peaks[i].fwhm, 1);
     }
@@ -331,23 +273,21 @@ describe('Optimize sum of Gaussians', () => {
   it('minima peaks', () => {
     let peakList = [
       {
-        x: -0.5,
+        x: -0.52,
         y: 0,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
       {
-        x: 0.5,
+        x: 0.52,
         y: 0,
         fwhm: 0.08,
         shape: { kind: 'gaussian' } as Shape1D,
       },
     ];
     let modifiedData = data;
-    modifiedData.y.map((value) => 1 - value);
-    let result = optimize(modifiedData, peakList, {
-      shape: { kind: 'gaussian' },
-    });
+    modifiedData.y.map((value: number) => 1 - value);
+    let result = optimize(modifiedData, peakList);
     for (let i = 0; i < 2; i++) {
       let pFit = result.peaks[i];
       expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
@@ -357,74 +297,32 @@ describe('Optimize sum of Gaussians', () => {
   });
 });
 
-describe('Optimize 4 parameters of a linear combination of gaussian and lorentzians', () => {
+describe('Sum of Pseudo Voigts', () => {
   const peaks = [
     {
       x: 0,
       y: 0.001,
       fwhm: 0.31,
-      mu: (xFactor * nbPoints) / 10,
-      shape: { kind: 'pseudoVoigt' } as Shape1D,
+      shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
     },
     {
       x: 0,
       y: 0.001,
       fwhm: 0.31,
       mu: (xFactor * nbPoints) / 10,
-      shape: { kind: 'pseudoVoigt' } as Shape1D,
+      shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
     },
   ];
-  const data: DataXY = generateSpectrum(peaks, {
+  const peaksGenerator = [
+    { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
+    { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
+  ];
+  const data: DataXY = generateSpectrum(peaksGenerator, {
     generator: {
-      from: -5,
-      to: 5,
-      nbPoints: 600,
+      from: -1,
+      to: 1,
+      nbPoints: 101,
     },
-  });
-
-  it('group of two GL', () => {
-    let pTrue = [
-      0,
-      0.001,
-      0.31,
-      (xFactor * nbPoints) / 10,
-      0,
-      0.001,
-      0.31,
-      (xFactor * nbPoints) / 10,
-    ];
-    let func = getSumOfShapes(peaks);
-    let yData = func(pTrue);
-    let result = optimize(
-      { x, y: x.map((i) => yData(i)) },
-      [
-        {
-          x: 0.1,
-          y: 0.0009,
-          fwhm: (xFactor * nbPoints) / 6,
-          shape: { kind: 'pseudoVoigt' } as Shape1D,
-        },
-        {
-          x: 0.1,
-          y: 0.0009,
-          fwhm: (xFactor * nbPoints) / 6,
-          shape: { kind: 'pseudoVoigt' } as Shape1D,
-        },
-      ],
-      {
-        shape: { kind: 'pseudoVoigt' },
-        optimization: {
-          kind: 'lm',
-          options: { maxIterations: 300, damping: 0.5, errorTolerance: 1e-8 },
-        },
-      },
-    );
-    for (let i = 0; i < 2; i++) {
-      let pFit = result.peaks[i];
-      expect(pFit.x).toBeCloseTo(pTrue[4 * i], 3);
-      expect(pFit.y).toBeCloseTo(pTrue[4 * i + 1], 3);
-      expect(pFit.fwhm).toBeCloseTo(pTrue[4 * i + 2], 2);
-    }
   });
 
   it('positive maxima peaks', () => {
@@ -433,19 +331,16 @@ describe('Optimize 4 parameters of a linear combination of gaussian and lorentzi
         x: 0.001,
         y: 0.0009,
         fwhm: (xFactor * nbPoints) / 6,
-        mu: (xFactor * nbPoints) / 10,
-        shape: { kind: 'pseudoVoigt' } as Shape1D,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
       },
       {
         x: 0.001,
         y: 0.0009,
         fwhm: (xFactor * nbPoints) / 6,
-        mu: (xFactor * nbPoints) / 10,
-        shape: { kind: 'pseudoVoigt' } as Shape1D,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
       },
     ];
     let result = optimize(data, peakList, {
-      shape: { kind: 'pseudoVoigt' },
       optimization: {
         kind: 'lm',
         options: { maxIterations: 500, damping: 0.5, errorTolerance: 1e-8 },
@@ -459,33 +354,50 @@ describe('Optimize 4 parameters of a linear combination of gaussian and lorentzi
     }
   });
 
-  it('Negative peaks', () => {
-    const peaksGenerator = [
-      { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
-      { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
-    ];
-    const data: DataXY = generateSpectrum(peaksGenerator, {
-      generator: {
-        from: -1,
-        to: 1,
-        nbPoints: 101,
-      },
-    });
-
+  it('negative maxima peaks', () => {
     let peakList = [
       {
         x: 0.001,
         y: -0.0009,
         fwhm: (xFactor * nbPoints) / 6,
-        mu: (xFactor * nbPoints) / 10,
-        shape: { kind: 'pseudoVoigt' } as Shape1D,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
       },
       {
         x: 0.001,
         y: -0.0009,
         fwhm: (xFactor * nbPoints) / 6,
-        mu: (xFactor * nbPoints) / 10,
-        shape: { kind: 'pseudoVoigt' } as Shape1D,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
+      },
+    ];
+    let modifiedData = data;
+    modifiedData.y.map((value: number) => value - 2);
+    let result = optimize(modifiedData, peakList, {
+      optimization: {
+        kind: 'lm',
+        options: { maxIterations: 300, damping: 0.5, errorTolerance: 1e-8 },
+      },
+    });
+    for (let i = 0; i < 2; i++) {
+      let pFit = result.peaks[i];
+      expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
+      expect(pFit.y).toBeCloseTo(peaks[i].y, 1);
+      expect(pFit.fwhm).toBeCloseTo(peaks[i].fwhm, 1);
+    }
+  });
+
+  it('Negative peaks', () => {
+    let peakList = [
+      {
+        x: 0.001,
+        y: -0.0009,
+        fwhm: (xFactor * nbPoints) / 6,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
+      },
+      {
+        x: 0.001,
+        y: -0.0009,
+        fwhm: (xFactor * nbPoints) / 6,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
       },
     ];
     let modifiedData = data;
@@ -505,31 +417,18 @@ describe('Optimize 4 parameters of a linear combination of gaussian and lorentzi
   });
 
   it('minima peaks', () => {
-    const peaksGenerator = [
-      { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
-      { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
-    ];
-    const data: DataXY = generateSpectrum(peaksGenerator, {
-      generator: {
-        from: -1,
-        to: 1,
-        nbPoints: 101,
-      },
-    });
     let peakList = [
       {
         x: 0.001,
         y: 0.0009,
         fwhm: (xFactor * nbPoints) / 6,
-        mu: (xFactor * nbPoints) / 10,
-        shape: { kind: 'pseudoVoigt' } as Shape1D,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
       },
       {
         x: 0.001,
         y: 0.0009,
         fwhm: (xFactor * nbPoints) / 6,
-        mu: (xFactor * nbPoints) / 10,
-        shape: { kind: 'pseudoVoigt' } as Shape1D,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
       },
     ];
     let modifiedData = data;
@@ -546,6 +445,132 @@ describe('Optimize 4 parameters of a linear combination of gaussian and lorentzi
       expect(pFit.x).toBeCloseTo(peaks[i].x, 1);
       expect(pFit.y).toBeCloseTo(peaks[i].y, 1);
       expect(pFit.fwhm).toBeCloseTo(peaks[i].fwhm, 1);
+    }
+  });
+});
+
+describe('Sum of a mix of distributions', () => {
+  it('group of two GL', () => {
+    let peaks = [
+      {
+        x: 0,
+        y: 0.001,
+        fwhm: 0.31,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D
+      },
+      {
+        x: 0,
+        y: 0.001,
+        fwhm: 0.31,
+        shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D
+      },
+      { x: -0.5, y: 0.001, fwhm: 0.31, shape: { kind: 'gaussian' } },
+      { x: 0.5, y: 0.001, fwhm: 0.31, shape: { kind: 'gaussian' } },
+      { x: -0.5, y: 0.001, fwhm: 0.31, shape: { kind: 'lorentzian' } },
+      { x: 0.5, y: 0.001, fwhm: 0.31, shape: { kind: 'lorentzian' } },
+    ];
+
+    let pTrue = [
+      0,
+      0.001,
+      0.31,
+      (xFactor * nbPoints) / 10,
+      0,
+      0.001,
+      0.31,
+      (xFactor * nbPoints) / 10,
+      -0.5,
+      0.001,
+      0.31,
+      0.5,
+      0.001,
+      0.31,
+      -0.5,
+      0.001,
+      0.31,
+      0.5,
+      0.001,
+      0.31,
+    ];
+
+    const peaksGenerator = [
+      { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
+      { x: 0, y: 0.001, fwhm: 0.31, mu: (xFactor * nbPoints) / 10 },
+      { x: -0.5, y: 0.001, fwhm: 0.31 },
+      { x: 0.5, y: 0.001, fwhm: 0.31 },
+      { x: -0.5, y: 0.001, fwhm: 0.31 },
+      { x: 0.5, y: 0.001, fwhm: 0.31 },
+    ];
+
+    const data: DataXY = generateSpectrum(peaksGenerator, {
+      generator: {
+        from: -1,
+        to: 1,
+        nbPoints: 101,
+      },
+    });
+
+    let result = optimize(
+      data,
+      [
+        {
+          x: 0.001,
+          y: 0.0009,
+          fwhm: (xFactor * nbPoints) / 8,
+          shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
+        },
+        {
+          x: 0.001,
+          y: 0.0009,
+          fwhm: (xFactor * nbPoints) / 8,
+          shape: {kind: 'pseudoVoigt', options: {mu: (xFactor * nbPoints) / 10 }} as Shape1D,
+        },
+        {
+          x: -0.52,
+          y: 0.0009,
+          fwhm: (xFactor * nbPoints) / 8,
+          shape: { kind: 'gaussian' } as Shape1D,
+        },
+        {
+          x: 0.52,
+          y: 0.0009,
+          fwhm: (xFactor * nbPoints) / 8,
+          shape: { kind: 'gaussian' } as Shape1D,
+        },
+        {
+          x: -0.52,
+          y: 0.0009,
+          fwhm: (xFactor * nbPoints) / 8,
+          shape: { kind: 'lorentzian' } as Shape1D,
+        },
+        {
+          x: 0.52,
+          y: 0.0009,
+          fwhm: (xFactor * nbPoints) / 8,
+          shape: { kind: 'lorentzian' } as Shape1D,
+        },
+      ],
+      {
+        optimization: {
+          kind: 'lm',
+          options: { maxIterations: 1000, errorTolerance: 1e-8 },
+        },
+      },
+    );
+    let index = 0;
+    for (let i = 0; i < 6; i++) {
+      let pFit = result.peaks[i];
+      if (peaks[i].shape.kind === 'pseudoVoigt') {
+        expect(pFit.x).toBeCloseTo(pTrue[index], 0);
+        expect(pFit.y).toBeCloseTo(pTrue[index + 1], 0);
+        expect(pFit.fwhm).toBeCloseTo(pTrue[index + 2], 0);
+        index += 4;
+      } else {
+        expect(pFit.x).toBeCloseTo(pTrue[index], 0);
+        expect(pFit.y).toBeCloseTo(pTrue[index + 1], 0);
+        expect(pFit.fwhm).toBeCloseTo(pTrue[index + 2], 0);
+        index += 3;
+      }
     }
   });
 });
